Function rf24_sys::root::RF24_getDynamicPayloadSize

pub unsafe extern "C" fn RF24_getDynamicPayloadSize(this: *mut RF24) -> u8

Get Dynamic Payload Size

For dynamic payloads, this pulls the size of the payload off the chip

@note Corrupt packets are now detected and flushed per the manufacturer. @code if(radio.available()){ if(radio.getDynamicPayloadSize() < 1){ // Corrupt payload has been flushed return; } radio.read(&data,sizeof(data)); } @endcode

@return Payload length of last-received dynamic payload