Function rf24_sys::root::RF24_openReadingPipe

pub unsafe extern "C" fn RF24_openReadingPipe(
    this: *mut RF24, 
    number: u8, 
    address: *const u8
)

Open a pipe for reading

Up to 6 pipes can be open for reading at once. Open all the required reading pipes, and then call startListening().

@see openWritingPipe @see setAddressWidth

@note Pipes 0 and 1 will store a full 5-byte address. Pipes 2-5 will technically only store a single byte, borrowing up to 4 additional bytes from pipe #1 per the assigned address width. @warning Pipes 1-5 should share the same address, except the first byte. Only the first byte in the array should be unique, e.g. @code uint8_t addresses[][6] = {"1Node","2Node"}; openReadingPipe(1,addresses[0]); openReadingPipe(2,addresses[1]); @endcode

@warning Pipe 0 is also used by the writing pipe. So if you open pipe 0 for reading, and then startListening(), it will overwrite the writing pipe. Ergo, do an openWritingPipe() again before write().

@param number Which pipe# to open, 0-5. @param address The 24, 32 or 40 bit address of the pipe to open.