Function rf24_sys::root::RF24_setPayloadSize

pub unsafe extern "C" fn RF24_setPayloadSize(this: *mut RF24, size: u8)

Set Static Payload Size

This implementation uses a pre-stablished fixed payload size for all transmissions. If this method is never called, the driver will always transmit the maximum payload size (32 bytes), no matter how much was sent to write().

@todo Implement variable-sized payloads feature

@param size The number of bytes in the payload