Function rf24_sys::root::RF24_txStandBy

pub unsafe extern "C" fn RF24_txStandBy(this: *mut RF24) -> bool

This function should be called as soon as transmission is finished to drop the radio back to STANDBY-I mode. If not issued, the radio will remain in STANDBY-II mode which, per the data sheet, is not a recommended operating mode.

@note When transmitting data in rapid succession, it is still recommended by the manufacturer to drop the radio out of TX or STANDBY-II mode if there is time enough between sends for the FIFOs to empty. This is not required if auto-ack is enabled.

Relies on built-in auto retry functionality.

@code Example (Partial blocking):

	radio.writeFast(&buf,32);
	radio.writeFast(&buf,32);
	radio.writeFast(&buf,32);  //Fills the FIFO buffers up
	bool ok = txStandBy();     //Returns 0 if failed. 1 if success.
			  				   //Blocks only until MAX_RT timeout or success. Data flushed on fail.

@endcode @see txStandBy(unsigned long timeout) @return True if transmission is successful