Function rf24_sys::root::RF24_whatHappened

pub unsafe extern "C" fn RF24_whatHappened(
    this: *mut RF24, 
    tx_ok: *mut bool, 
    tx_fail: *mut bool, 
    rx_ready: *mut bool
)

Call this when you get an interrupt to find out why

Tells you what caused the interrupt, and clears the state of interrupts.

@param[out] tx_ok The send was successful (TX_DS) @param[out] tx_fail The send failed, too many retries (MAX_RT) @param[out] rx_ready There is a message waiting to be read (RX_DS)