Function rf24_sys::root::RF24_write

pub unsafe extern "C" fn RF24_write(
    this: *mut RF24, 
    buf: *const c_void, 
    len: u8
) -> bool

Be sure to call openWritingPipe() first to set the destination of where to write to.

This blocks until the message is successfully acknowledged by the receiver or the timeout/retransmit maxima are reached. In the current configuration, the max delay here is 60-70ms.

The maximum size of data written is the fixed payload size, see getPayloadSize(). However, you can write less, and the remainder will just be filled with zeroes.

TX/RX/RT interrupt flags will be cleared every time write is called

@param buf Pointer to the data to be sent @param len Number of bytes to be sent

@code radio.stopListening(); radio.write(&data,sizeof(data)); @endcode @return True if the payload was delivered successfully and an ACK was received, or upon successfull transmission if auto-ack is disabled.