Function rf24_sys::root::RF24_writeAckPayload

pub unsafe extern "C" fn RF24_writeAckPayload(
    this: *mut RF24, 
    pipe: u8, 
    buf: *const c_void, 
    len: u8
)

Write an ack payload for the specified pipe

The next time a message is received on @p pipe, the data in @p buf will be sent back in the acknowledgement. @see enableAckPayload() @see enableDynamicPayloads() @warning Only three of these can be pending at any time as there are only 3 FIFO buffers.
Dynamic payloads must be enabled. @note Ack payloads are handled automatically by the radio chip when a payload is received. Users should generally write an ack payload as soon as startListening() is called, so one is available when a regular payload is received. @note Ack payloads are dynamic payloads. This only works on pipes 0&1 by default. Call enableDynamicPayloads() to enable on all pipes.

@param pipe Which pipe# (typically 1-5) will get this response. @param buf Pointer to data that is sent @param len Length of the data to send, up to 32 bytes max. Not affected by the static payload set by setPayloadSize().