Function rf24_sys::root::RF24_writeBlocking

pub unsafe extern "C" fn RF24_writeBlocking(
    this: *mut RF24, 
    buf: *const c_void, 
    len: u8, 
    timeout: u32
) -> bool

This function extends the auto-retry mechanism to any specified duration. It will not block until the 3 FIFO buffers are filled with data. If so the library will auto retry until a new payload is written or the user specified timeout period is reached. @warning It is important to never keep the nRF24L01 in TX mode and FIFO full for more than 4ms at a time. If the auto retransmit is enabled, the nRF24L01 is never in TX mode long enough to disobey this rule. Allow the FIFO to clear by issuing txStandBy() or ensure appropriate time between transmissions.

@code Example (Full blocking):

	radio.writeBlocking(&buf,32,1000); //Wait up to 1 second to write 1 payload to the buffers
	txStandBy(1000);     			   //Wait up to 1 second for the payload to send. Return 1 if ok, 0 if failed.
			  				   		   //Blocks only until user timeout or success. Data flushed on fail.

@endcode @note If used from within an interrupt, the interrupt should be disabled until completion, and sei(); called to enable millis(). @see txStandBy() @see write() @see writeFast()

@param buf Pointer to the data to be sent @param len Number of bytes to be sent @param timeout User defined timeout in milliseconds. @return True if the payload was loaded into the buffer successfully false if not