Function rf24_sys::root::RF24_writeFast

pub unsafe extern "C" fn RF24_writeFast(
    this: *mut RF24, 
    buf: *const c_void, 
    len: u8
) -> bool

This will not block until the 3 FIFO buffers are filled with data. Once the FIFOs are full, writeFast will simply wait for success or timeout, and return 1 or 0 respectively. From a user perspective, just keep trying to send the same data. The library will keep auto retrying the current payload using the built in functionality. @warning It is important to never keep the nRF24L01 in TX mode and FIFO full for more than 4ms at a time. If the auto retransmit is enabled, the nRF24L01 is never in TX mode long enough to disobey this rule. Allow the FIFO to clear by issuing txStandBy() or ensure appropriate time between transmissions.

@code Example (Partial blocking):

	radio.writeFast(&buf,32);  // Writes 1 payload to the buffers
	txStandBy();     		   // Returns 0 if failed. 1 if success. Blocks only until MAX_RT timeout or success. Data flushed on fail.

	radio.writeFast(&buf,32);  // Writes 1 payload to the buffers
	txStandBy(1000);		   // Using extended timeouts, returns 1 if success. Retries failed payloads for 1 seconds before returning 0.

@endcode

@see txStandBy() @see write() @see writeBlocking()

@param buf Pointer to the data to be sent @param len Number of bytes to be sent @return True if the payload was delivered successfully false if not