Struct rf24_sys::root::RF24

#[repr(C)]
pub struct RF24 {
    pub vtable_: *const RF24__bindgen_vtable,
    pub spi: SPI,
    pub ce_pin: u16,
    pub csn_pin: u16,
    pub spi_speed: u16,
    pub spi_rxbuff: [u8; 33],
    pub spi_txbuff: [u8; 33],
    pub p_variant: bool,
    pub payload_size: u8,
    pub dynamic_payloads_enabled: bool,
    pub pipe0_reading_address: [u8; 5],
    pub addr_width: u8,
    pub failureDetected: bool,
    pub txDelay: u32,
    pub csDelay: u32,
}

Driver for nRF24L01(+) 2.4GHz Wireless Transceiver

Fields
vtable_: *const RF24__bindgen_vtablespi: SPIce_pin: u16

< "Chip Enable" pin, activates the RX or TX role

csn_pin: u16

< SPI Chip select

spi_speed: u16

< SPI Bus Speed

spi_rxbuff: [u8; 33]spi_txbuff: [u8; 33]p_variant: boolpayload_size: u8

< Fixed size of payloads

dynamic_payloads_enabled: bool

< Whether dynamic payloads are enabled.

pipe0_reading_address: [u8; 5]

< Last address set on pipe 0 for reading.

addr_width: u8

< The address width to use - 3,4 or 5 bytes.

failureDetected: booltxDelay: u32

The driver will delay for this duration when stopListening() is called

When responding to payloads, faster devices like ARM(RPi) are much faster than Arduino:

Arduino sends data to RPi, switches to RX mode

The RPi receives the data, switches to TX mode and sends before the Arduino radio is in RX mode

If AutoACK is disabled, this can be set as low as 0. If AA/ESB enabled, set to 100uS minimum on RPi

@warning If set to 0, ensure 130uS delay after stopListening() and before any sends

csDelay: u32

On all devices but Linux and ATTiny, a small delay is added to the CSN toggling function

This is intended to minimise the speed of SPI polling due to radio commands

If using interrupts or timed requests, this can be set to 0 Default:5

Methods

impl RF24

pub unsafe fn begin(&mut self) -> bool

pub unsafe fn isChipConnected(&mut self) -> bool

pub unsafe fn startListening(&mut self)

pub unsafe fn stopListening(&mut self)

pub unsafe fn available(&mut self) -> bool

pub unsafe fn read(&mut self, buf: *mut c_void, len: u8)

pub unsafe fn write(&mut self, buf: *const c_void, len: u8) -> bool

pub unsafe fn openWritingPipe(&mut self, address: *const u8)

pub unsafe fn openReadingPipe(&mut self, number: u8, address: *const u8)

pub unsafe fn printDetails(&mut self)

pub unsafe fn available1(&mut self, pipe_num: *mut u8) -> bool

pub unsafe fn rxFifoFull(&mut self) -> bool

pub unsafe fn powerDown(&mut self)

pub unsafe fn powerUp(&mut self)

pub unsafe fn write1(
&mut self,
buf: *const c_void,
len: u8,
multicast: bool
) -> bool

pub unsafe fn writeFast(&mut self, buf: *const c_void, len: u8) -> bool

pub unsafe fn writeFast1(
&mut self,
buf: *const c_void,
len: u8,
multicast: bool
) -> bool

pub unsafe fn writeBlocking(
&mut self,
buf: *const c_void,
len: u8,
timeout: u32
) -> bool

pub unsafe fn txStandBy(&mut self) -> bool

pub unsafe fn txStandBy1(&mut self, timeout: u32, startTx: bool) -> bool

pub unsafe fn writeAckPayload(&mut self, pipe: u8, buf: *const c_void, len: u8)

pub unsafe fn isAckPayloadAvailable(&mut self) -> bool

pub unsafe fn whatHappened(
&mut self,
tx_ok: *mut bool,
tx_fail: *mut bool,
rx_ready: *mut bool
)

pub unsafe fn startFastWrite(
&mut self,
buf: *const c_void,
len: u8,
multicast: bool,
startTx: bool
)

pub unsafe fn startWrite(
&mut self,
buf: *const c_void,
len: u8,
multicast: bool
)

pub unsafe fn reUseTX(&mut self)

pub unsafe fn flush_tx(&mut self) -> u8

pub unsafe fn testCarrier(&mut self) -> bool

pub unsafe fn testRPD(&mut self) -> bool

pub unsafe fn closeReadingPipe(&mut self, pipe: u8)

pub unsafe fn setAddressWidth(&mut self, a_width: u8)

pub unsafe fn setRetries(&mut self, delay: u8, count: u8)

pub unsafe fn setChannel(&mut self, channel: u8)

pub unsafe fn getChannel(&mut self) -> u8

pub unsafe fn setPayloadSize(&mut self, size: u8)

pub unsafe fn getPayloadSize(&mut self) -> u8

pub unsafe fn getDynamicPayloadSize(&mut self) -> u8

pub unsafe fn enableAckPayload(&mut self)

pub unsafe fn enableDynamicPayloads(&mut self)

pub unsafe fn disableDynamicPayloads(&mut self)

pub unsafe fn enableDynamicAck(&mut self)

pub unsafe fn isPVariant(&mut self) -> bool

pub unsafe fn setAutoAck(&mut self, enable: bool)

pub unsafe fn setAutoAck1(&mut self, pipe: u8, enable: bool)

pub unsafe fn setPALevel(&mut self, level: u8)

pub unsafe fn getPALevel(&mut self) -> u8

pub unsafe fn setDataRate(&mut self, speed: rf24_datarate_e) -> bool

pub unsafe fn getDataRate(&mut self) -> rf24_datarate_e

pub unsafe fn setCRCLength(&mut self, length: rf24_crclength_e)

pub unsafe fn getCRCLength(&mut self) -> rf24_crclength_e

pub unsafe fn disableCRC(&mut self)

pub unsafe fn maskIRQ(&mut self, tx_ok: bool, tx_fail: bool, rx_ready: bool)

pub unsafe fn openReadingPipe1(&mut self, number: u8, address: u64)

pub unsafe fn openWritingPipe1(&mut self, address: u64)

pub unsafe fn flush_rx(&mut self) -> u8

pub unsafe fn new(_cepin: u16, _cspin: u16) -> Self

pub unsafe fn new1(_cepin: u16, _cspin: u16, spispeed: u32) -> Self